Fix igraph vcount return type and R package name
Change vapply FUN.VALUE from integer(1) to numeric(1) to match
igraph::vcount returning double in recent versions. Rename R package
from dynamic_multiplex to dynamicmultiplex.

+++ r_code/R/fit_multilayer_identity_ties.R
@@ -32,7 +32,7 @@ fit_multilayer_identity_ties <- function(layers,
   )
 
   n_nodes <- igraph::vcount(graph_layers[[1]])
-  if (!all(vapply(graph_layers, igraph::vcount, integer(1)) == n_nodes)) {
+  if (!all(vapply(graph_layers, igraph::vcount, numeric(1)) == n_nodes)) {
     stop("All layers must have the same number of nodes for identity ties.", call. = FALSE)
   }
